+++ nextshot.sh
@@ -114,7 +114,8 @@ tray_menu() {
     rm -f "$_TRAY_FIFO"; mkfifo "$_TRAY_FIFO" && exec 3<> "$_TRAY_FIFO"
 
     yad --notification --listen --no-middle --command="nextshot -a" <&3 &
-    echo $! > "$_RUNTIME_DIR/traymenu.pid"
+    local traypid=$!
+    echo $traypid > "$_RUNTIME_DIR/traymenu.pid"
 
     echo "menu:\
 Open Nextcloud      ! xdg-open $files_url !emblem-web||\
@@ -122,7 +123,7 @@ Capture area        ! nextshot -a         !window-maximize-symbolic|\
 Capture window      ! nextshot -w         !window-new|\
 Capture full screen ! nextshot -f         !view-fullscreen-symbolic||\
 Paste from Clipboard! nextshot -p         !edit-paste-symbolic||\
-Quit Nextshot       !quit                 !gtk-quit" >&3
+Quit Nextshot       ! kill $traypid       !gtk-quit" >&3
 
     echo "icon:camera-photo-symbolic" >&3
     echo "tooltip:Nextshot" >&3