+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+* CLI prompt for renaming screenshots when running interactively
 
 
 ## [1.1.1] - 2019-04-04