+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-* CLI prompt for renaming screenshots when running interactively
+* Ability to force environment with `NEXTSHOT_ENV` var or `--env` flag
+
+### Changed
+* Filename prompt now uses CLI instead of Yad when running interactively
 
 ### Fixed
 * Quit item in tray menu didn't actually quit